Refactor webhook and message interfaces to common namespace and improve project structure
