--- LineBridge/Enums/Webhook/EventEnums.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-
-namespace LineBridge.Enums.Webhook
-{
-    public class EventEnums
-    {
-        /// <summary>
-        /// イベントタイプ
-        /// </summary>
-        [JsonConverter(typeof(StringEnumConverter), true)]
-        public enum EventType
-        {
-            /// <summary>
-            /// メッセージイベント
-            /// </summary>
-            Message
-        }
-    }
-}
