+++ LineBridge/Interfaces/Webhook/INagiyuWebhook.cs
@@ -4,6 +4,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
+using LineBridge.Common.Interfaces.Webhook;
+
 namespace LineBridge.Interfaces.Webhook
 {
     public interface INagiyuWebhook : IWebhookBase
