--- LineBridge/Interfaces/Webhook/IWebhookBase.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LineBridge.Interfaces.Webhook
-{
-    public interface IWebhookBase
-    {
-        /// <summary>
-        /// Webhook イベントを処理する
-        /// </summary>
-        /// <param name="requestBody">リクエストボディ</param>
-        Task HandleWebhookEvent(string requestBody);
-    }
-}
