--- LineBridge/Models/Webhook/Events/Message/MessageEvent.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using LineBridge.Models.Webhook.Events.Message.Objects;
-
-namespace LineBridge.Models.Webhook.Events.Message
-{
-    public class MessageEvent<T> : EventBase where T : ObjectBase
-    {
-        /// <summary>
-        /// 応答メッセージを送る際に使用する応答トークン
-        /// </summary>
-        public string ReplyToken { get; set; }
-
-        /// <summary>
-        /// メッセージの内容を含むオブジェクト
-        /// </summary>
-        public T Message { get; set; }
-    }
-}
