+++ NagiyuLine.sln
@@ -17,7 +17,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DynamoDBAccessor.Tests", "D
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LineBridge", "LineBridge\LineBridge.csproj", "{0CE06334-46CA-45E6-849B-BB00F72BB3A0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenAIConnect.Common", "OpenAIConnect.Common\OpenAIConnect.Common.csproj", "{50010262-13BD-458F-A28A-3A660D83D291}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenAIConnect.Common", "OpenAIConnect.Common\OpenAIConnect.Common.csproj", "{50010262-13BD-458F-A28A-3A660D83D291}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LineBridge.Common", "LineBridge.Common\LineBridge.Common.csproj", "{25CFC659-61A4-40C1-896E-9EB6AA21E954}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LineBridge.Core", "LineBridge.Core\LineBridge.Core.csproj", "{6EF5D597-1283-4FDD-97B9-F238378293D2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -57,6 +61,14 @@ Global
 		{50010262-13BD-458F-A28A-3A660D83D291}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{50010262-13BD-458F-A28A-3A660D83D291}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{50010262-13BD-458F-A28A-3A660D83D291}.Release|Any CPU.Build.0 = Release|Any CPU
+		{25CFC659-61A4-40C1-896E-9EB6AA21E954}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{25CFC659-61A4-40C1-896E-9EB6AA21E954}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{25CFC659-61A4-40C1-896E-9EB6AA21E954}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{25CFC659-61A4-40C1-896E-9EB6AA21E954}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6EF5D597-1283-4FDD-97B9-F238378293D2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6EF5D597-1283-4FDD-97B9-F238378293D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6EF5D597-1283-4FDD-97B9-F238378293D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6EF5D597-1283-4FDD-97B9-F238378293D2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
